+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## Unreleased
+
+### Added
+
+- `--migration-start-ref` option to `odoo-helper test` command, that allows
+  to specify the starting point for migration test.
+
+---
+
 ## Release 1.0.1 (2023-01-10)
 
 ### Fixed